Staging

+++ src/main/java/org/gaung/wiwokdetok/fondasikehidupan/model/BookLocation.java
@@ -1,7 +1,16 @@
 package org.gaung.wiwokdetok.fondasikehidupan.model;
 
-import jakarta.persistence.*;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.locationtech.jts.geom.Point;
 
 @Entity
@@ -13,7 +22,7 @@ public class BookLocation {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(name = "location_name", nullable = false)
     private String locationName;
