+++ src/main/java/org/gaung/wiwokdetok/fondasikehidupan/model/Genre.java
@@ -1,7 +1,14 @@
 package org.gaung.wiwokdetok.fondasikehidupan.model;
 
-import jakarta.persistence.*;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Table(name = "genre")
@@ -12,7 +19,7 @@ public class Genre {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(length = 15, nullable = false, unique = true)
     private String genre;
